:truck: api: move scoring certification service from lib to src
